Fixed issue #496: app export now skips files that do not exist

+++ SexyContent/ImportExport/ZipExport.cs
@@ -90,7 +90,8 @@ namespace ToSic.SexyContent.ImportExport
                 if (!Directory.Exists(portalFilePath))
                     Directory.CreateDirectory(portalFilePath);
                 
-                File.Copy(file.PhysicalPath, Path.Combine(portalFilesDirectory.FullName, file.RelativePath.Replace('/', '\\')));
+                if (File.Exists(file.PhysicalPath))
+                    File.Copy(file.PhysicalPath, Path.Combine(portalFilesDirectory.FullName, file.RelativePath.Replace('/', '\\')));
             }
             
             // Save export xml
